+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Adds `ngram` function
 - Adds `is_empty` and `is_nonempty` functions
 - Provide a raw view of the query in the RequestResult
+- Adds backward support for old references
 
 ## 2.0.0 (March 19, 2018)
 - Adds support for @query type (user defined functions)