+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 2.0.3-SNAPSHOT
+- Bug fix: Typescript functions that receive lambda should support lambdas created
+  by `Lambda()` function.
 
 ## 2.0.2 (March 28, 2018)
 - Bug fix: functions with optional scope parameter were failing when scope was