+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 2.0.3-SNAPSHOT
+## 2.1.0 (July 25, 2018)
 - Bug fix: Typescript functions that receive lambda should support lambdas created
   by `Lambda()` function.
 - Adds `to_string`, `to_number`, `to_time`, and `to_date` functions